refactor: resolve #293 - reduce interfaces.ts from 692 to under 500 lines
Closes #293
